Calibrate Features Values Using a Reference Sample

Calibrates features abundances using a reference sample, either by re-calibrating absolute concentrations using known concentrations in the reference samples or by calculating relative ratios to reference samples.

Two calibration approaches are supported:

Absolute calibration (absolute_calibration = TRUE): Recalibrates features to concentrations using known reference sample values.

Requirements:

analyte_id and sample_id must be defined in analysis metadata

Known analyte concentrations must be defined for the reference sample in QC concentration metadata

The function:

Overwrites original values (intensity, norm_intensity, or conc) with recalibrated values

Stores original values in feature_[VARIABLE]_beforecal, where [VARIABLE] is conc, norm_intensity, or intensity

Can process conc, norm_intensity, or intensity data

Always outputs concentration (conc) in concentration unit defined for the analytes in the reference sample in the QC concentration metadata.

When multiple analyses of the same reference sample are present, their concentrations are summarized using summarize_fun ("mean" or "median", default: "mean").

Options for undefined analyte concentrations for the reference sample (undefined_conc_handling):

"original": Keep original values (meaning the non-calibrated values will be returned)

"na": Set undefined features to NA

"error": Raise error for undefined features An error will be raised if no concentrations are defined for any features.

To export:

Use save_dataset_csv() with variable = "[VARIABLE]

For MiDAR XLSX report, use save_report_xlsx() also with variable = "[VARIABLE] to include the re-calibrated values of the filtered dataset in the report.

Set [VARIABLE] = conc, norm_intensity, or intensity to export re-calibrated values, or conc_beforecal, norm_intensity_beforecal, or intensity_beforecal to export the original values.

Relative calibration (absolute_calibration = FALSE): Expresses sample features values as ratios relative to corresponding features in the reference sample.

Applies uniformly to all features

Works with conc, norm_intensity, or intensity data

Stores results in feature_[VARIABLE]_normalized, where [VARIABLE] is conc, norm_intensity, or intensity

To export:

Use save_dataset_csv() with variable = "feature_[VARIABLE]_normalized"

For MiDAR XLSX report, use save_report_xlsx() with same variable setting as for save_dataset_csv() to include the normalized values of the filtered dataset in the report. Unfiltered normalized feature values will be included by default in the report if present

Usage

calibrate_by_reference(
  data,
  variable,
  reference_sample_id,
  absolute_calibration,
  summarize_fun = "mean",
  undefined_conc_action = NULL
)

Arguments

data

A MidarExperiment object containing the metabolomics data to be normalized

variable

Character string indicating which data type to calibrate Must be one of: "intensity", "norm_intensity", or "conc"

reference_sample_id

Character vector specifying the sample ID(s) to use as reference(s) or standards

absolute_calibration

Logical indicating whether to perform absolute calibration using known concentrations of the reference sample (TRUE) or relative calibration (FALSE).

summarize_fun

Either "mean" or "median". If absolute_calibration = TRUE, this function is used to summarize the reference sample concentrations across analyses of specified reference_sample_id. Default is "mean".

undefined_conc_action

Character string specifying how to handle features without defined concentrations in reference samples when absolute_calibration = TRUE. Must be one of: "original" (keep original values), "na" (set to NA), or "error". Default is "keep".

Value

A MidarExperiment object with calibrated data

See also

Examples

## ---- pkgdown = FALSE ------------------------------------------------
if (FALSE) { # \dontrun{
  # Relative calibration
  mexp <- calibrate_by_reference(
    data = mexp,
    variable = "conc",
    reference_sample_id = "NIST-SRM1950",
    absolute_calibration = TRUE,
    summarize_fun = "mean",
    undefined_conc_action = "original"
  )

  # Export relative calibration results
  save_dataset_csv(mexp, "mexp_calibrated.csv", variable = "conc")

  # Export non-calibrated results
  save_dataset_csv(mexp, "mexp_calibrated.csv", variable = "conc_beforecal")

  # Export relative calibration results in the MiDAR XLSX report
  save_report_xlsx(mexp, "report.xlsx", variable = "conc")

  # Absolute calibration example
  mexp <- calibrate_by_reference(
    data = mexp,
    variable = "conc",
    reference_sample_id = "NIST-SRM1950",
    absolute_calibration = FALSE
  )

  # Export relative calibration results
  save_dataset_csv(mexp, "mexp_calibrated.csv", variable = "conc_normalized")

  # Export relative calibration results in the MiDAR XLSX report
  save_report_xlsx(mexp, "report.xlsx", variable = "conc_normalized")
} # }