Import response curves metadata

Imports response curve metadata (annotation) from a preloaded data frame or tibble via the data argument, or from data from a file (CSV or Excel) via the path argument. The analysis metadata must contain following columns: analysis_id, curve_id, analyzed_amount and analyzed_amount_unit.

Usage

import_metadata_responsecurves(
  data = NULL,
  table = NULL,
  path = NULL,
  sheet = NULL,
  ignore_warnings = FALSE
)

Arguments

data

A MidarExperiment object

table

A data frame or tibble with response curve metadata. If path is also provided, an error will be raised.

path

A character string specifying the path to a CSV (.csv) or Excel (.xlsx) file. If table is also provided, an error will be raised.

sheet

Defines the sheet name in case an Excel file is provided.

ignore_warnings

Ignore warnings from data validation and proceed with importing metadata

Value

An updated MidarExperiment object