Plasma Lipidomics Peak Areas Dataset

This is an demo dataset used for examples in this of from reporocessed dataset containing raw peak areas and analytical metadata of 499 analyses with 503 features from a plasma lipidomics analysis published in Tan et al, ATVB, 2022. The dataset is used to demonstrate

Usage

lipidomics_dataset

Format

A MidarExperiment object:

dataset_orig

A tibble with the original peak data.

dataset

A tibble with annotated data.

annot_analyses

A tibble with analyses annotations.

annot_features

A tibble with features annotations.

annot_batches

A tibble with batch annotations.

...