Plot Overall QC Filtering Summary

This function generates a summary of the feature QC filtering process, visualizing the number of features that passed or failed the various QC criteria. It includes a Venn diagram showing the features excluded due to different filtering criteria such as signal-to-blank ratios, CV thresholds, and linearity. The criteria are applied hierarchically, meaning a feature must pass all lower-tier filters before being considered for failure on higher-tier filters.

Usage

plot_qc_summary_overall(
  data = NULL,
  include_qualifier = FALSE,
  include_istd = FALSE,
  with_venn_diag = TRUE,
  user_defined_keeper = FALSE,
  font_base_size = 8
)

Arguments

data

MidarExperiment object

include_qualifier

Whether to include qualifier features in the plot. Default is FALSE.

include_istd

Whether to include internal standard features in the plot. Default is FALSE.

with_venn_diag

Whether to include a Venn diagram summarizing the features excluded due to different QC criteria. Default is TRUE.

user_defined_keeper

Whether to retain user-specified features that were not selected by the QC filters. Default is FALSE.

font_base_size

The base font size for the plot. Default is 8.

Value

A ggplot2 object showing the feature QC filtering summary with or without a Venn diagram.

Details

The QC filtering process follows a hierarchical structure, where features are first evaluated against lower-level filters such as signal-to-blank ratios and limit of detection (LOD). Only features that pass these basic criteria are then subjected to higher-level filters like the coefficient of variation (CV) or linear regression results. A feature will only fail a higher-level filter (such as CV or R-squared) if it has passed all previous lower-level filters. This ensures that features are evaluated progressively, starting from fundamental quality checks up to more stringent filtering criteria.