Get FC and p values from a t test contrast

Get FC and p values from a t test contrast

Usage

volcano_plot(
  d_FC,
  p_adjust,
  sig_FC_min,
  sig_p_value_min,
  symmetric_x,
  x_min,
  x_max,
  point_size,
  point_transparency,
  scale_factor,
  scale_factor_species_label,
  squared = TRUE,
  axistick = 1
)

Arguments

d_FC

data.frame with FEATURE_NAME, p and log2FC

p_adjust

calculate and use FDR

sig_FC_min

|FC| treshold

sig_p_value_min

P value treshold

symmetric_x

make x axis symetric around 0

x_min

x min

x_max

x max

point_size

point size

point_transparency

point alpha

scale_factor

scale factor for fonts

scale_factor_species_label

scale factor point labels

squared

make plot a square

axistick

axis ticks

Value

tibble with p values, fdr and log2FC